Patrick's Old/requirements.txt
# Core
torch>=2.1.0
transformers>=4.40.0
datasets>=2.19.0
evaluate>=0.4.1
scikit-learn>=1.4.1
numpy>=1.26.4
scipy>=1.12.0
pandas>=2.2.2
tqdm>=4.66.4
uvicorn>=0.30.0
flask>=3.0.0
jinja2>=3.1.4
python-dotenv>=1.0.1
waitress

# PEFT (LoRA)
peft>=0.11.1

# SpaCy + SciSpaCy + linkers
spacy>=3.7.4
scispacy>=0.5.4
# models to install separately (large):
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_core_sci_scibert-0.5.4.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bc5cdr_md-0.5.4.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_jnlpba_md-0.5.4.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_craft_md-0.5.4.tar.gz
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.4/en_ner_bionlp13cg_md-0.5.4.tar.gz

sentence-transformers>=2.7.0

# Visualization
cython
networkx>=3.3
matplotlib>=3.9.2

# Streamlit GUI
streamlit>=1.35.0
openai>=1.30.0
